- All donation number is now genereted from JPA Repository ( count () method)

+++ src/main/java/pl/coderslab/charity/web/controllers/HomeController.java
@@ -30,7 +30,7 @@ public class HomeController {
     }
 
     @ModelAttribute (name = "allDonationsNumber")
-    private Integer allDonationNUmber(){
+    private Long allDonationNUmber(){
         return donationServices.allDonationNumber();
     }
